+++ mex/common/backend_api/models.py
@@ -1,41 +1,23 @@
-from typing import Annotated
+from typing import Annotated, Generic, TypeVar
 
 from pydantic import Field, TypeAdapter
 
-from mex.common.models import (
-    AnyExtractedModel,
-    AnyMergedModel,
-    AnyPreviewModel,
-    AnyRuleSetResponse,
-    BaseModel,
-)
+from mex.common.models import AnyMergedModel, AnyRuleSetResponse, BaseModel
 from mex.common.types import Identifier
 
-
-class ExtractedItemsRequest(BaseModel):
-    """Request model for a list of extracted items."""
-
-    items: list[AnyExtractedModel]
+T = TypeVar("T")
 
 
-class ExtractedItemsResponse(BaseModel):
-    """Response model for a list of extracted items including a total count."""
+class ItemsContainer(BaseModel, Generic[T]):
+    """Generic container that contains items."""
 
-    items: list[AnyExtractedModel]
-    total: int
-
-
-class MergedItemsResponse(BaseModel):
-    """Response model for a list of merged items including a total count."""
-
-    items: list[AnyMergedModel]
-    total: int
+    items: list[T]
 
 
-class PreviewItemsResponse(BaseModel):
-    """Response model for a list of preview items including a total count."""
+class PaginatedItemsContainer(BaseModel, Generic[T]):
+    """Generic container that contains items and has a total item count."""
 
-    items: list[AnyPreviewModel]
+    items: list[T]
     total: int